refactor: Move truncate_content to action_execution_server
- Move truncate_content function to action_execution_server.py
- Remove truncation from event.py and agent_controller.py
- Centralize all truncation at the server level
